Account for Veterans Day
In order to allow for Veteran's Day (which I forgot about because last year it was on a Friday) I have decided we will have a field trip to the Career Fair on Oct 23 instead of our regular class schedule. This will make sprints wonky until after Veteran's Day 11/11 and it will result in 8 sprints instead of the revised 9. What can I say, Welcoming Changing Requirements isn't easy.